Move RazorViewEngineOptionsSetup to internal namespace
